// src/enums.ts
export enum MEILISEARCH_ERROR_CODES {
  INDEX_NOT_FOUND = "index_not_found",
  API_KEY_NOT_FOUND = "API_KEY_NOT_FOUND",
  HOST_NOT_FOUND = "HOST_NOT_FOUND",
  SERVICE_NOT_FOUND = "SERVICE_NOT_FOUND",
  SUBSCRIBER_SETTING_NOT_FOUND = "SUBSCRIBER_SETTING_NOT_FOUND",
  ENTITY_NOT_FOUND = "ENTITY_NOT_FOUND",
}